chore: add extendable store
