+++ packages/preact/preact.ts
@@ -11,6 +11,6 @@ export { useActions } from "./src/hooks/useActions"
 export { useDecoratedSearchResults } from "./src/hooks/useDecoratedSearchResults"
 export { useNostoAppState } from "./src/hooks/useNostoAppState"
 export { usePersonalization } from "./src/hooks/usePersonalization"
-export { createStore, type State, type Store } from "./src/store"
+export { createStore, defaultState, type State, type Store } from "./src/store"
 export { StoreContext } from "./src/storeContext"
-export { mockStore } from "./test/mocks/mocks"
+export { createExtendableStore, type ExtendedStore } from "./src/storeExtensions"
