+++ packages/preact/src/store.tsx
@@ -1,9 +1,6 @@
 import type { SearchQuery, SearchResult } from "@nosto/nosto-js/client"
-import { deepFreeze } from "@utils/deepFreeze"
-import { deepMerge } from "@utils/deepMerge"
-import { isEqual } from "@utils/isEqual"
 
-import { PageType } from "./api/types"
+import { createExtendableStore } from "./storeExtensions"
 
 /**
  * Preact state includes all changing variables of the app.
@@ -20,10 +17,6 @@ export interface State {
    */
   query: SearchQuery
   /**
-   * Displays which page type it is - category | search
-   */
-  pageType: PageType | undefined
-  /**
    * Current search response that includes found products, keywords and other results.
    */
   response: SearchResult
@@ -39,7 +32,6 @@ export interface State {
 
 export const defaultState: State = {
   loading: true,
-  pageType: undefined,
   query: {
     query: ""
   },
@@ -49,55 +41,14 @@ export const defaultState: State = {
   initialized: false
 }
 
+/**
+ * Create a standard Nosto store to manage the state of the app.
+ *
+ * If you use Nosto providers, you don't need to create a store manually.
+ * @param overrides Initial state overrides.
+ */
 export function createStore(overrides: Partial<State> = {}) {
-  const changeCallbacks: Map<(piece: never) => void, (state: State) => void> = new Map()
-  let state: State = overrides ? deepMerge(defaultState, overrides) : defaultState
-  const initialState = deepFreeze(state)
-
-  function setState(newState: (prevState: State) => State) {
-    state = newState(state)
-    for (const callback of changeCallbacks.values()) {
-      callback(state)
-    }
-  }
-
-  function updateState(newState: Partial<State>) {
-    setState(prevState => {
-      return { ...prevState, ...newState }
-    })
-  }
-
-  function getState() {
-    return state
-  }
-
-  function getInitialState() {
-    return initialState
-  }
-
-  function onChange<T>(selector: (state: State) => T, callback: (piece: T) => void) {
-    let lastValue: T | undefined
-
-    changeCallbacks.set(callback, newState => {
-      const piece = selector(newState)
-      if (!isEqual(piece, lastValue)) {
-        lastValue = piece
-        callback(piece)
-      }
-    })
-  }
-
-  function clearOnChange<T>(callback: (piece: T) => void) {
-    changeCallbacks.delete(callback)
-  }
-
-  return {
-    updateState,
-    getState,
-    getInitialState,
-    onChange,
-    clearOnChange
-  }
+  return createExtendableStore(defaultState, overrides)
 }
 
 export type Store = ReturnType<typeof createStore>
